requirements.txt
streamlit
openai
python-dotenv
pillow
tiktoken
